Finished Lab9
Added Authorization for every action in the app, with 4 roles: Unregistered User, User, Editor and Admin

+++ ArticlesApp/Controllers/CommentsController.cs
@@ -1,5 +1,7 @@
 ﻿using ArticlesApp.Data;
 using ArticlesApp.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -9,20 +11,33 @@ namespace ArticlesApp.Controllers
     {
 
         private readonly ApplicationDbContext db;
+		private readonly UserManager<ApplicationUser> userManager;
+		private readonly RoleManager<IdentityRole> roleManager;
 
-		public CommentsController(ApplicationDbContext db)
-		{
-			this.db = db;
-		}
+        public CommentsController(ApplicationDbContext context, UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
+        {
+            db = context;
+            this.userManager = userManager;
+            this.roleManager = roleManager;
+        }
 
-		[HttpGet]
-		public IActionResult Delete(int id)
+        [HttpGet]
+		[Authorize(Roles = "User,Editor,Admin")]
+        public IActionResult Delete(int id)
 		{
 			try
 			{
 				Comment comment = db.Comments.Find(id);
 				int? articleId = comment.ArticleId;
-				db.Comments.Remove(comment);
+
+				if(comment.UserId != userManager.GetUserId(User) && !User.IsInRole("Admin"))
+                {
+                    TempData["error"] = "Nu aveti dreptul sa stergeti acest comentariu!";
+                    return RedirectToAction("Show", "Articles", new { id = articleId });
+                }
+
+
+                db.Comments.Remove(comment);
 				db.SaveChanges();
                 TempData["message"] = "Comentariul a fost sters!";
                 return RedirectToAction("Show", "Articles", new { id = articleId });
